chore: refactor / relocate utils
